refactor: elimitate deep dependencies into parser
