+++ packages/openapi-ts/src/utils/enum.ts
@@ -1,5 +1,5 @@
 import type { Enum, Model } from '../openApi';
-import { ensureValidTypeScriptJavaScriptIdentifier } from '../openApi/common/parser/sanitize';
+import { ensureValidTypeScriptJavaScriptIdentifier } from '../openApi';
 import { unescapeName } from './escape';
 import { sort } from './sort';
 import { unique } from './unique';
